+++ docs/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gate ACE behind workflow eligibility to prevent spurious affects on manually-triaged,
   untriaged, and legacy flaws (OSIDB-4972, OSIDB-5307, OSIDB-5312)
 
+### Removed
+- delete deprecated alerts (OSIDB-5337)
+
 ## [5.14.0] - 2026-08-05
 ### Changed
 - set potential-rejection label as additional MANUAL workflow condition